Working migration to REST API

+++ components/CardGrid/CardGrid.tsx
@@ -4,84 +4,77 @@ import { BiGitPullRequest } from "react-icons/bi";
 import { VscIssues, VscRepoForked } from "react-icons/vsc";
 import { BarPlot } from "../../components/BarPlot";
 import { CardLink } from "../../components/CardLink";
-import { Repo } from "../../models";
+// import { Repo } from "../../models";
 import styles from "./CardGrid.module.css";
-import { mutate } from "swr";
 
 type Props = {
-  // repoData: Repo[];
-  list: string[];
-  repoData: any;
+  session: any;
+  data: any;
   loading: boolean;
+  error: boolean;
+  mutate: any;
 };
 
-const CardGrid = ({ list, repoData, loading }: Props) => {
-  if (repoData === "loading") {
-    // if (loading === "loading") {
+const CardGrid = ({ session, data, loading, error, mutate }: Props) => {
+  if (loading) {
     return <h2>Loading...</h2>;
   }
 
-  if (repoData === "error") {
-    // if (loading === "error") {
-    // console.error(error);
-    return null;
+  if (error) {
+    console.error(error);
+    return <h2>Failed to load data</h2>;
   }
 
-  const handleDelete = async (id: string, list: string[]) => {
-    const updatedList = list.filter((item: any) => {
-      return item.node_id != id;
+  const handleDelete = async (id: string, data: object[]) => {
+    const updatedList = data.filter((element) => {
+      return element.id != id;
     });
 
-    const response = await fetch(`/api/repositories?id=${id}`, {
-      method: "DELETE",
-    });
+    console.log("updatedList", updatedList);
 
-    if (!response.ok) {
-      throw new Error(`Error: ${response.status}`);
+    // optimistic UI
+    mutate(updatedList, false);
+
+    if (session) {
+      const response = await fetch(`/api/repos?id=${id}`, {
+        method: "DELETE",
+      });
+      if (!response.ok) {
+        throw new Error(`Error: ${response.status}`);
+      }
+      mutate();
     }
-    mutate("/api/repositories");
   };
 
-  if (
-    !repoData ||
-    repoData.length === 0 ||
-    repoData.error === "Could not resolve to a node with the global id of ''"
-  ) {
-    return null;
-  }
-
-  console.log("list", list);
-  console.log("data", repoData);
-
-  const repoCards = repoData?.data.nodes.map((repo: Repo, index: any) => {
+  const repoCards = data.map((repo) => {
     return (
       <div className={styles.repocard} key={repo.id}>
         <div className={styles.close}>
-          <button onClick={() => handleDelete(repo.id, list)}>Remove</button>
+          <button onClick={() => handleDelete(repo.id, data)}>Remove</button>
         </div>
         <CardLink>
-          <h2>{repo.nameWithOwner}</h2>
+          <h2>
+            {repo.owner}/{repo.repo}
+          </h2>
           <h3>{repo.description}</h3>
           <div className={styles.content}>
             <div className={styles.list}>
               <p>
-                <AiFillStar /> stars: <strong>{repo.stargazerCount}</strong>
+                <AiFillStar /> stars: <strong>{repo.starGazers}</strong>
               </p>
               <p>
-                <VscIssues /> issues (open):{" "}
-                <strong>{repo.openIssues.totalCount}</strong>
+                <VscIssues /> issues (open): <strong>{repo.issues}</strong>
               </p>
               <p>
-                <VscRepoForked /> forks: <strong>{repo.forkCount}</strong>
+                <VscRepoForked /> forks: <strong>{repo.forks}</strong>
               </p>
               <p>
-                <BiGitPullRequest /> pull requests (open):{" "}
-                <strong>{repo.pullRequests.totalCount}</strong>
+                <BiGitPullRequest /> watchers: <strong>{repo.watchers}</strong>
               </p>
             </div>
             <div className={styles.plot}>
               <div className={styles.chart}>
-                <BarPlot name={repo.nameWithOwner} />
+                <BarPlot name={repo.repo} owner={repo.owner} />
               </div>
             </div>
           </div>
